requirements.txt
# Requirements

pyyaml-5.3.1
PIL

